+++ caravan/tests/commands/test_start.py
@@ -1,78 +1,171 @@
+import unittest
+import json
+
+import httpretty
+from abduct import captured, out, err
+
 from caravan.commands.start import Command
 
 
-def run_and_get_mock(mocker, args):
-    orig_args = ['PROG', '-d', 'DOMAIN', '-n', 'NAME', '-v', 'VERSION',
-                 '-i', 'ID']
-    args = orig_args + args
-    mocker.patch('sys.argv', args)
-
-    run_swf = mocker.patch('caravan.commands.start.run_swf_command')
-    run_swf.return_value = {'runId': 'RUNID'}
-
-    Command.main()
-
-    return run_swf
+httpretty.HTTPretty.allow_net_connect = False
 
 
-def test_minimal(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=[])
-
-    run_swf.assert_called_once_with(
-        'start_workflow_execution',
-        domain='DOMAIN',
-        workflowId='ID',
-        workflowType={'name': 'NAME', 'version': 'VERSION'},
-        input=None,
-        taskList=None,
-        tagList=None,
-        executionStartToCloseTimeout=None,
-        taskStartToCloseTimeout=None,
-        childPolicy=None,
-        lambdaRole=None,
-        )
-
-    out, _ = capsys.readouterr()
-    assert 'Execution started.' in out
-    assert 'RunId: RUNID' in out
+def httpretty_register():
+    headers = {
+        'x-amzn-RequestId': 'd68969c7-3f0d-11e1-9b11-7182192d0b57',
+        }
+    body = '{"runId": "1e536162-f1ea-48b0-85f3-aade88eef2f7"}'
+    httpretty.register_uri(httpretty.POST,
+                           "https://swf.us-east-1.amazonaws.com/",
+                           content_type='application/json',
+                           adding_headers=headers,
+                           body=body)
 
 
-def test_task_list(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['-t', 'TASKLISTNAME'])
-    assert run_swf.call_args[1]['taskList'] == {'name': 'TASKLISTNAME'}
+def run_command(additional_args):
+    httpretty_register()
+
+    args = ['-d', 'DOMAIN', '-n', 'NAME', '-v', 'VERSION', '-i', 'ID']
+    args += additional_args
+    with captured(out(), err()) as (stdout, stderr):
+        Command.main(args=args)
+
+    return stdout, stderr
 
 
-def test_tag_list_1(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--tag', 'TAG1'])
-    assert run_swf.call_args[1]['tagList'] == ['TAG1']
+class Test(unittest.TestCase):
 
+    @httpretty.activate
+    def test_nominal(self):
+        httpretty_register()
+        stdout, stderr = run_command([])
+        request = httpretty.last_request()
 
-def test_tag_list_2(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--tag', 'TAG1 TAG2'])
-    assert run_swf.call_args[1]['tagList'] == ['TAG1', 'TAG2']
+        self.assertEqual(request.headers.get('x-amz-target'),
+                         'SimpleWorkflowService.StartWorkflowExecution')
 
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            }
+        self.assertEqual(json.loads(request.body), expected)
 
-def test_input(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--input', 'INPUT'])
-    assert run_swf.call_args[1]['input'] == 'INPUT'
+        self.assertIn('Execution started.', stdout.getvalue())
+        self.assertIn('1e536162-f1ea-48b0-85f3-aade88eef2f7',
+                      stdout.getvalue())
 
+    @httpretty.activate
+    def test_parameter_task_list(self):
+        httpretty_register()
+        run_command(['-t', 'TASKLISTNAME'])
+        request = httpretty.last_request()
 
-def test_execution_timeout(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--execution-timeout', '60'])
-    assert run_swf.call_args[1]['executionStartToCloseTimeout'] == '60'
+        expected = {
+            "domain": "DOMAIN",
+            "taskList": {"name": "TASKLISTNAME"},
+            "workflowId": "ID",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            }
+        self.assertEqual(json.loads(request.body), expected)
 
+    @httpretty.activate
+    def test_parameter_tag(self):
+        httpretty_register()
+        run_command(['--tag', 'TAG'])
+        request = httpretty.last_request()
 
-def test_task_timeout(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--task-timeout', '60'])
-    assert run_swf.call_args[1]['taskStartToCloseTimeout'] == '60'
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "tagList": ["TAG"],
+            }
+        self.assertEqual(json.loads(request.body), expected)
 
+    @httpretty.activate
+    def test_parameter_tag_two(self):
+        httpretty_register()
+        run_command(['--tag', 'TAG1 TAG2'])
+        request = httpretty.last_request()
 
-def test_child_policy(mocker, capsys):
-    test_policy = Command.CHILD_POLICIES[0]
-    run_swf = run_and_get_mock(mocker, args=['--child-policy', test_policy])
-    assert run_swf.call_args[1]['childPolicy'] == test_policy
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "tagList": ["TAG1", "TAG2"]
+            }
+        self.assertEqual(json.loads(request.body), expected)
 
+    @httpretty.activate
+    def test_parameter_input(self):
+        httpretty_register()
+        run_command(['--input', 'INPUT'])
+        request = httpretty.last_request()
 
-def test_lambda_role(mocker, capsys):
-    run_swf = run_and_get_mock(mocker, args=['--lambda-role', 'ROLE'])
-    assert run_swf.call_args[1]['lambdaRole'] == 'ROLE'
+        expected = {
+            "input": "INPUT",
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID"
+            }
+        self.assertEqual(json.loads(request.body), expected)
+
+    @httpretty.activate
+    def test_parameter_execution_timeout(self):
+        httpretty_register()
+        run_command(['--execution-timeout', '60'])
+        request = httpretty.last_request()
+
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "executionStartToCloseTimeout": "60",
+            }
+        self.assertEqual(json.loads(request.body), expected)
+
+    @httpretty.activate
+    def test_parameter_task_timeout(self):
+        httpretty_register()
+        run_command(['--task-timeout', '60'])
+        request = httpretty.last_request()
+
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "taskStartToCloseTimeout": "60",
+            }
+        self.assertEqual(json.loads(request.body), expected)
+
+    @httpretty.activate
+    def test_parameter_child_policy(self):
+        test_policy = Command.CHILD_POLICIES[0]
+        httpretty_register()
+        run_command(['--child-policy', test_policy])
+        request = httpretty.last_request()
+
+        Command.CHILD_POLICIES[0]
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "childPolicy": test_policy,
+            }
+        self.assertEqual(json.loads(request.body), expected)
+
+    @httpretty.activate
+    def test_parameter_lambda_role(self):
+        httpretty_register()
+        run_command(['--lambda-role', 'ROLE'])
+        request = httpretty.last_request()
+
+        Command.CHILD_POLICIES[0]
+        expected = {
+            "domain": "DOMAIN",
+            "workflowType": {"version": "VERSION", "name": "NAME"},
+            "workflowId": "ID",
+            "lambdaRole": 'ROLE',
+            }
+        self.assertEqual(json.loads(request.body), expected)